+++ cordapp/src/main/kotlin/net/corda/plugins/CordappPlugin.kt
@@ -1,6 +1,10 @@
 package net.corda.plugins
 
+import net.corda.plugins.SigningOptions.Companion.DEFAULT_KEYSTORE
+import net.corda.plugins.SigningOptions.Companion.DEFAULT_KEYSTORE_EXTENSION
+import net.corda.plugins.SigningOptions.Companion.DEFAULT_KEYSTORE_FILE
 import net.corda.plugins.Utils.Companion.compareVersions
+import net.corda.plugins.Utils.Companion.createTempFileFromResource
 import org.gradle.api.GradleException
 import org.gradle.api.InvalidUserDataException
 import org.gradle.api.Plugin
@@ -9,6 +13,7 @@ import org.gradle.api.artifacts.Configuration
 import org.gradle.api.artifacts.Dependency
 import org.gradle.jvm.tasks.Jar
 import java.io.File
+import java.nio.file.Paths
 
 /**
  * The Cordapp plugin will turn a project into a cordapp project which builds cordapp JARs with the correct format
@@ -76,15 +81,39 @@ class CordappPlugin : Plugin<Project> {
         }
         jarTask.dependsOn(task)
 
-        val signTask = project.task("signCordappJar")
-        signTask.doLast {
-            if (cordapp.signing.enabled) {
-                val options = cordapp.signing.options.toSignJarOptionsMap()
-                options["jar"] = project.tasks.getByName("jar").outputs.files.singleFile.toPath().toString()
-                project.ant.invokeMethod("signjar", options)
+        val signTask = project.task("signJar") { it.doLast { sign(project, cordapp.signing) } }
+        jarTask.finalizedBy(signTask)
+    }
+
+    private fun sign(project: Project, signing: Signing) {
+        if (!signing.enabled) {
+            project.logger.info("CorDapp JAR signing is disabled, the CorDapp's contracts will not use signature constraints.")
+            return
+        }
+        val options = signing.options.toSignJarOptionsMap()
+        if (signing.options.hasDefaultOptions()) {
+            project.logger.info("CorDapp JAR signing with the default Corda development key, suitable for Corda running in development mode only.")
+            val keyStorePath = createTempFileFromResource(DEFAULT_KEYSTORE, DEFAULT_KEYSTORE_FILE, DEFAULT_KEYSTORE_EXTENSION)
+            options["keystore"] = keyStorePath.toString()
+        }
+
+        val path = project.tasks.getByName("jar").outputs.files.singleFile.toPath()
+        options["jar"] = path.toString()
+
+        try {
+            project.ant.invokeMethod("signjar", options)
+        } catch (e: Exception) {
+            // Not adding error message as it's always meaningless, logs with --INFO level contain more insights
+            throw InvalidUserDataException("Exception while signing ${path.fileName}, " +
+                    "ensure the 'cordapp.signing.options' entry contains correct keyStore configuration, " +
+                    "or disable signing by 'cordapp.signing.enabled false'. " +
+                    if (project.logger.isInfoEnabled || project.logger.isDebugEnabled) "Search for 'ant:signjar' in log output."
+                    else "Run with --info or --debug option and search for 'ant:signjar' in log output. ", e)
+        } finally {
+            if (signing.options.hasDefaultOptions()) {
+                Paths.get(options["keystore"]).toFile().delete()
             }
         }
-        jarTask.finalizedBy(signTask)
     }
 
     private fun getDirectNonCordaDependencies(project: Project): Set<File> {